worker/transcriber/requirements.txt
openai-whisper
kafka-python
python-dotenv